♻️ refactor folder staking -> tasks_staking
